Combine binary and hh count logic into one unified pipeline
Now the files need only specify a tibble of signals to produce, and
parameters of how to produce those signals. aggregate.R now does a
signle batch of geo and day aggregations, and then calculates all
indicators for that geo-day. This prevents duplicate aggregation and
streamlines the code.

This will also make it easier to add new indicators -- simply add new
rows to the appropriate tibble, and they will be instantly reported in
the new aggregate output.