Correct individual response inclusion criteria
1. The count of how many questions were answered was incorrect, since
there are now additional columns in the data frame. Count again and
adjust the threshold.

2. The old pipeline does not *remove* invalid ZIPs or those from
counties with <100 population. It intended to blank out ZIPs with small
populations and preserve invalid ZIPs; however, the blanking was never
actually implemented. Implement it here, while keeping invalid ZIPs, and
adjust tests to match.