set default num days to export to all since export start date
Since BigQuery tables are currently not partitioned by date, each
query processes and bills for all rows in the table regardless of filters
applied (date and country, at the moment). To take advantage of this,
this pipeline will pull all dates from the specified start date to the
current date by default. This setting should be updated to a shorter
date window (~14 days seems reasonable) if/when the tables are converted
to "partitioned" format.