update GS params to specify credential values
GS params previously set path to BigQuery credentials file, which
doesn't work well with the Ansible vault used in production
(specifically, this doesn't allow for encrypting the API key). This
switches from a path to putting BQ credentials directly in the params
file, for both local and Ansible params files. Tests and code are
updated accordingly.