simplify code_activities to process one wave at a time
`code_activities` is called in `load_response_one` so it only sees one
wave of data at a time. Thus, we don't anticipate needing to handle both
a C13 and a C13b (new "indoors" version of C13) together. Remove some
logic that allows this, and replace with an assert verifying that we are
in fact handling only one wave. Update tests to reflect new behavior.