Add v0.2 COVID survey variables and contingency table definitions
* Add v0.2 grouping and indicator variables

In contingency_variables.R, add all v0.2 grouping and indicator
variables, and reorganize the variables to be in the same order as
they appear in the requirements spreadsheet for easy checking.

* Add v0.2 contingency tables

In contingency_indicators.R, add definitions for all v0.2 indicators
for the requested contingency tables, e.g. nation, nation/age, etc.