Use col spec from input df to read output file
readr's column guessing procedure only uses the first 1000 lines, by
default, of a file to guess variable type for each column. If a column
is completely missing for the first 1000 lines, it is read in as a
logical which causes parsing failures if the column contains non-boolean
values later, outside the type guessing range.

This happens when reading in output files if an indicator was newly
added. To correctly specify these, use the column specification from the
input file/s. All columns included in input files are at least partially
non-missing and sorted alphabetically (indepdendent of missingness), so
we should always see non-missing values in the first 1000 lines.