process reports for auto-backfill only if missing signals
Previously, `fetch_new_reports` and CSV saving would happen in the
auto-backfill chunk regardless of if COVIDCcast metadata was missing any
signals. Because of a bug in `fetch_listing`, this would cause the
entire report history to be processed on every pipeline run.

This moves the report processing so that it is only triggered if any
signals that we want are not found in metadata.