short-circuit `fetch_listing` if no new reports found
The `if params['indicator']['reports'] == 'new':` chunk properly
produces an empty `keep` list if no new reports are available in a
run. But empty `keep` lists are always ignored in the return
statement of the function, which returns `listing` (the full set of
reports) instead. The result is that when no new reports are
available (e.g. on the weekend or with runs in quick succession)
the pipeline process the entire report history.